LayoutTests/imported/w3c/web-platform-tests/selection/shadow-dom/tentative/Selection-getComposedRanges-dom-mutations-removal_mode=open-expected.txt

slotted content 2

FAIL Range is fully in shadow tree. Removing shadow host collapses composed StaticRange. Note it does not update previously returned composed StaticRange. Type error
FAIL Range is fully in shadow tree. Removing parent of shadow host collapses composed StaticRange. Type error
PASS Range is in light DOM. Removing startContainer rescopes new composed range to its parent.
FAIL Range is across shadow trees. Replacing shadowRoot content rescopes new composed range to the shadowRoot. Type error
FAIL Range is across shadow trees. Removing ancestor shadow host rescopes composed range end to parent. Type error
FAIL Range is between two light slotted contents. Removing start container rescopes to its parent in light tree. Type error
